Move import options into CreateWalletImportScene
