+++ eslint.config.mjs
@@ -239,9 +239,6 @@ export default [
       'src/components/scenes/CreateWalletAccountSelectScene.tsx',
       'src/components/scenes/CreateWalletAccountSetupScene.tsx',
 
-      'src/components/scenes/CreateWalletImportOptionsScene.tsx',
-      'src/components/scenes/CreateWalletImportScene.tsx',
-
       'src/components/scenes/CurrencyNotificationScene.tsx',
       'src/components/scenes/DefaultFiatSettingScene.tsx',
       'src/components/scenes/DuressModeHowToScene.tsx',
